refactor(dashboard): extract observability skill folders
Summary:
- move the terminal dashboard source into dashboard/terminal-ui
- move the observability router, controller, and presenter into
  dashboard/http-api
- extract throughput and sparkline helpers into
  SymphonyElixir.Dashboard.Metrics and wire Mix to compile dashboard/*/lib
- add skill descriptors and focused metrics coverage for the extracted layout

Rationale:
- align the observability code with the issue's self-contained dashboard
  folder structure without changing the existing public module surface
- keep the shared pub/sub interface stable while separating terminal UI,
  metrics, and HTTP API responsibilities for future iteration
- preserve the repository's existing coverage policy for extracted
  observability helper modules while keeping behavior validated

Tests:
- mix test test/symphony_elixir/dashboard_metrics_test.exs test/symphony_elixir/status_dashboard_snapshot_test.exs test/symphony_elixir/orchestrator_status_test.exs test/symphony_elixir/extensions_test.exs test/symphony_elixir/observability_pubsub_test.exs
- make -C elixir all
