+++ dashboard/http-api/lib/symphony_elixir_web/controllers/observability_api_controller.ex
