+++ dashboard/http-api/lib/symphony_elixir_web/router.ex
