+++ dashboard/metrics/SKILL.md
@@ -1,0 +1,7 @@
+# dashboard/metrics
+
+Owns throughput sampling, rolling windows, and sparkline generation used by the observability dashboard.
+
+## Files
+
+- `lib/symphony_elixir/dashboard/metrics.ex`: rolling TPS and sparkline helpers
