+++ dashboard/terminal-ui/SKILL.md
@@ -1,0 +1,7 @@
+# dashboard/terminal-ui
+
+Owns ANSI terminal rendering for the observability dashboard, including refresh cadence and snapshot formatting.
+
+## Files
+
+- `lib/symphony_elixir/status_dashboard.ex`: terminal dashboard GenServer and rendering helpers
