+++ dashboard/terminal-ui/lib/symphony_elixir/status_dashboard.ex
@@ -7,14 +7,11 @@ defmodule SymphonyElixir.StatusDashboard do
   require Logger
 
   alias SymphonyElixir.{Config, HttpServer}
+  alias SymphonyElixir.Dashboard.Metrics
   alias SymphonyElixir.Orchestrator
   alias SymphonyElixirWeb.ObservabilityPubSub
 
   @minimum_idle_rerender_ms 1_000
-  @throughput_window_ms 5_000
-  @throughput_graph_window_ms 10 * 60 * 1000
-  @throughput_graph_columns 24
-  @sparkline_blocks ["▁", "▂", "▃", "▄", "▅", "▆", "▇", "█"]
   @running_id_width 8
   @running_stage_width 14
   @running_pid_width 8
@@ -198,7 +195,7 @@ defmodule SymphonyElixir.StatusDashboard do
     current_tokens = snapshot_total_tokens(snapshot_data)
 
     {tps_second, tps} =
-      throttled_tps(
+      Metrics.throttled_tps(
         state.last_tps_second,
         state.last_tps_value,
         now_ms,
@@ -319,13 +316,13 @@ defmodule SymphonyElixir.StatusDashboard do
              rate_limits: Map.get(snapshot, :rate_limits),
              polling: Map.get(snapshot, :polling)
            }},
-          update_token_samples(token_samples, now_ms, total_tokens)
+          Metrics.update_token_samples(token_samples, now_ms, total_tokens)
         }
 
       :error ->
         {
           :error,
-          prune_samples(token_samples, now_ms)
+          Metrics.prune_samples(token_samples, now_ms)
         }
     end
   end
@@ -474,59 +471,15 @@ defmodule SymphonyElixir.StatusDashboard do
     ])
   end
 
-  defp update_token_samples(samples, now_ms, total_tokens) do
-    prune_graph_samples([{now_ms, total_tokens} | samples], now_ms)
-  end
-
-  defp prune_samples(samples, now_ms) do
-    min_timestamp = now_ms - @throughput_window_ms
-    Enum.filter(samples, fn {timestamp, _} -> timestamp >= min_timestamp end)
-  end
-
-  defp prune_graph_samples(samples, now_ms) do
-    min_timestamp = now_ms - max(@throughput_window_ms, @throughput_graph_window_ms)
-    Enum.filter(samples, fn {timestamp, _} -> timestamp >= min_timestamp end)
-  end
-
   @doc false
   @spec rolling_tps([{integer(), integer()}], integer(), integer()) :: float()
-  def rolling_tps(samples, now_ms, current_tokens) do
-    samples = [{now_ms, current_tokens} | samples]
-    samples = prune_samples(samples, now_ms)
-
-    case samples do
-      [] ->
-        0.0
-
-      [_one] ->
-        0.0
-
-      _ ->
-        first = List.last(samples)
-        {start_ms, start_tokens} = first
-        elapsed_ms = now_ms - start_ms
-        delta_tokens = max(0, current_tokens - start_tokens)
-
-        if elapsed_ms <= 0 do
-          0.0
-        else
-          delta_tokens / (elapsed_ms / 1000.0)
-        end
-    end
-  end
+  def rolling_tps(samples, now_ms, current_tokens), do: Metrics.rolling_tps(samples, now_ms, current_tokens)
 
   @doc false
   @spec throttled_tps(integer() | nil, float() | nil, integer(), [{integer(), integer()}], integer()) ::
           {integer(), float()}
-  def throttled_tps(last_second, last_value, now_ms, token_samples, current_tokens) do
-    second = div(now_ms, 1000)
-
-    if is_integer(last_second) and last_second == second and is_number(last_value) do
-      {second, last_value}
-    else
-      {second, rolling_tps(token_samples, now_ms, current_tokens)}
-    end
-  end
+  def throttled_tps(last_second, last_value, now_ms, token_samples, current_tokens),
+    do: Metrics.throttled_tps(last_second, last_value, now_ms, token_samples, current_tokens)
 
   @doc false
   @spec format_timestamp_for_test(DateTime.t()) :: String.t()
@@ -648,7 +601,7 @@ defmodule SymphonyElixir.StatusDashboard do
 
   @doc false
   @spec tps_graph_for_test([{integer(), integer()}], integer(), integer()) :: String.t()
-  def tps_graph_for_test(samples, now_ms, current_tokens), do: tps_graph(samples, now_ms, current_tokens)
+  def tps_graph_for_test(samples, now_ms, current_tokens), do: Metrics.tps_graph(samples, now_ms, current_tokens)
 
   defp format_retry_rows(retrying) do
     if retrying == [] do
@@ -871,65 +824,6 @@ defmodule SymphonyElixir.StatusDashboard do
     |> Integer.to_string()
     |> group_thousands()
   end
-
-  defp tps_graph(samples, now_ms, current_tokens) do
-    bucket_ms = div(@throughput_graph_window_ms, @throughput_graph_columns)
-    active_bucket_start = div(now_ms, bucket_ms) * bucket_ms
-    graph_window_start = active_bucket_start - (@throughput_graph_columns - 1) * bucket_ms
-
-    rates =
-      [{now_ms, current_tokens} | samples]
-      |> prune_graph_samples(now_ms)
-      |> Enum.sort_by(&elem(&1, 0))
-      |> Enum.chunk_every(2, 1, :discard)
-      |> Enum.map(fn [{start_ms, start_tokens}, {end_ms, end_tokens}] ->
-        elapsed_ms = end_ms - start_ms
-        delta_tokens = max(0, end_tokens - start_tokens)
-        tps = if elapsed_ms <= 0, do: 0.0, else: delta_tokens / (elapsed_ms / 1000.0)
-        {end_ms, tps}
-      end)
-
-    bucketed_tps =
-      0..(@throughput_graph_columns - 1)
-      |> Enum.map(fn bucket_idx ->
-        bucket_start = graph_window_start + bucket_idx * bucket_ms
-        bucket_end = bucket_start + bucket_ms
-        last_bucket? = bucket_idx == @throughput_graph_columns - 1
-
-        values =
-          rates
-          |> Enum.filter(fn {timestamp, _tps} ->
-            in_bucket?(timestamp, bucket_start, bucket_end, last_bucket?)
-          end)
-          |> Enum.map(fn {_timestamp, tps} -> tps end)
-
-        if values == [] do
-          0.0
-        else
-          Enum.sum(values) / length(values)
-        end
-      end)
-
-    max_tps = Enum.max(bucketed_tps, fn -> 0.0 end)
-
-    bucketed_tps
-    |> Enum.map_join(fn value ->
-      index =
-        if max_tps <= 0 do
-          0
-        else
-          round(value / max_tps * (length(@sparkline_blocks) - 1))
-        end
-
-      Enum.at(@sparkline_blocks, index, "▁")
-    end)
-  end
-
-  defp in_bucket?(timestamp, bucket_start, bucket_end, true),
-    do: timestamp >= bucket_start and timestamp <= bucket_end
-
-  defp in_bucket?(timestamp, bucket_start, bucket_end, false),
-    do: timestamp >= bucket_start and timestamp < bucket_end
 
   defp format_rate_limits(nil), do: colorize("unavailable", @ansi_gray)
